requirements.txt
py-trello==0.10.0
terminaltables==3.1.0
colorama==0.3.9
